[GH-8327] Deprecate EntityManagerHelper for a provider abstraction. (#8524)
* cli config

* [GH-8327] Deprecate EntityManager HelperSet for a provider abstraction.

* Housekeeping: phpcs

* [GH-8327] Refactor tests towards use of SingleManagerProvider instead of HelperSet.

* [GH-8327] Refactor tests towards use of SingleManagerProvider instead of HelperSet.

* [GH-8327] Refactor tests towards use of SingleManagerProvider instead of HelperSet.

* Housekeeping: cs

* Update tests/Doctrine/Tests/ORM/Tools/Console/ConsoleRunnerTest.php


* [GH-8327] Change option from entity-manager to em for consistency with DoctrineBundle.

* Add final to new methods and classes


* [GH-8327] Bugfix: phpstan detected stricter type checks needed.

* phpcs

+++ lib/Doctrine/ORM/Tools/Console/Command/ClearCache/MetadataCommand.php
@@ -21,21 +21,26 @@
 namespace Doctrine\ORM\Tools\Console\Command\ClearCache;
 
 use Doctrine\Common\Cache\ApcCache;
+use Doctrine\Common\Cache\ClearableCache;
+use Doctrine\Common\Cache\FlushableCache;
 use Doctrine\Common\Cache\XcacheCache;
+use Doctrine\ORM\Tools\Console\Command\AbstractEntityManagerCommand;
 use InvalidArgumentException;
 use LogicException;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
+
+use function get_class;
+use function sprintf;
 
 /**
  * Command to clear the metadata cache of the various cache drivers.
  *
  * @link    www.doctrine-project.org
  */
-class MetadataCommand extends Command
+class MetadataCommand extends AbstractEntityManagerCommand
 {
     /**
      * {@inheritdoc}
@@ -44,6 +49,7 @@ class MetadataCommand extends Command
     {
         $this->setName('orm:clear-cache:metadata')
              ->setDescription('Clear all metadata cache of the various cache drivers')
+             ->addOption('em', null, InputOption::VALUE_REQUIRED, 'Name of the entity manager to operate on', 'default')
              ->addOption('flush', null, InputOption::VALUE_NONE, 'If defined, cache entries will be flushed instead of deleted/invalidated.')
              ->setHelp(<<<EOT
 The <info>%command.name%</info> command is meant to clear the metadata cache of associated Entity Manager.
@@ -72,7 +78,7 @@ EOT
     {
         $ui = new SymfonyStyle($input, $output);
 
-        $em          = $this->getHelper('em')->getEntityManager();
+        $em          = $this->getEntityManager($input);
         $cacheDriver = $em->getConfiguration()->getMetadataCacheImpl();
 
         if (! $cacheDriver) {
@@ -87,12 +93,26 @@ EOT
             throw new LogicException('Cannot clear XCache Cache from Console, its shared in the Webserver memory and not accessible from the CLI.');
         }
 
+        if (! ($cacheDriver instanceof ClearableCache)) {
+            throw new LogicException(sprintf(
+                'Can only clear cache when ClearableCache interface is implemented, %s does not implement.',
+                get_class($cacheDriver)
+            ));
+        }
+
         $ui->comment('Clearing <info>all</info> Metadata cache entries');
 
         $result  = $cacheDriver->deleteAll();
         $message = $result ? 'Successfully deleted cache entries.' : 'No cache entries were deleted.';
 
         if ($input->getOption('flush') === true) {
+            if (! ($cacheDriver instanceof FlushableCache)) {
+                throw new LogicException(sprintf(
+                    'Can only clear cache when FlushableCache interface is implemented, %s does not implement.',
+                    get_class($cacheDriver)
+                ));
+            }
+
             $result  = $cacheDriver->flushAll();
             $message = $result ? 'Successfully flushed cache entries.' : $message;
         }
